+++ app/src/main/java/util/common/NewRecipeListener.java
@@ -3,7 +3,7 @@ package util.common;
 import models.entities.Ingredient;
 
 public interface NewRecipeListener {
-    public void addIngredientToRecipe(Ingredient ingredient, Integer quantity);
+    public void addIngredientToRecipe(Ingredient ingredient, Float quantity);
 
     public void removeIngredientFromRecipe(Ingredient ingredient);
 }
